backend/requirements.txt
asgiref==3.2.10
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.2
django-cors-headers==3.5.0
django-rest-multiple-models==2.1.3
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
drf-writable-nested==0.6.2
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
riotwatcher==3.1.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
zipp==3.4.0
